question_generation/gui/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\archive\research\intuitive-physics\dataset\gui\dataset_viewer.py: 10
PyQt5 
PyQt5-sip == 12.8.0


# D:\archive\research\intuitive-physics\dataset\gui\dataset_viewer.py: 8
python-vlc
